requirements.txt
langchain
langchain_community
gradio
openai
googlesearch-python
python-docx
google-search-results
wikipedia
faiss-cpu
sentence_transformers
tiktoken
PyPDF2
python-dotenv
scikit-image
opencv-python
langchain_openai
gradio_tools
requests
beautifulsoup4